Add basic multiplayer using WebRTC/PeerJS.
Create a "server" peer that calculates the game state based on user
input and sends the state to the clients/players every frame. The
players/clients simply send their input state (where the mouse is and
what keys are pressed) to the server, and update the scene to reflect
the game state received from the server.